fix bool and None comparison in tests; move fixture to conftest.py
