+++ core/tests/conftest.py
@@ -4,20 +4,19 @@ import shutil
 from typing import Any
 from typing import Generator
 
-import pytest
+from qdrant_client import QdrantClient
+
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
+import pytest
 
 from cat.db.database import Database
-
 import cat.utils as utils
-
-from qdrant_client import QdrantClient
 from cat.memory.vector_memory import VectorMemory
 from cat.mad_hatter.plugin import Plugin
-
 from cat.main import cheshire_cat_api
+from tests.utils import create_mock_plugin_zip
 
 
 # substitute classes' methods where necessary for testing purposes
@@ -95,3 +94,31 @@ def client(app: FastAPI, monkeypatch) -> Generator[TestClient, Any, None]:
 
     with TestClient(app) as client:
         yield client
+
+
+# This fixture is useful to write tests in which
+#   a plugin was just uploaded via http.
+#   It wraps any test function having `just_installed_plugin` as an argument
+@pytest.fixture()
+def just_installed_plugin(client):
+    ### executed before each test function
+
+    # create zip file with a plugin
+    zip_path = create_mock_plugin_zip(flat=True)
+    zip_file_name = zip_path.split("/")[-1]  # mock_plugin.zip in tests/mocks folder
+
+    # upload plugin via endpoint
+    with open(zip_path, "rb") as f:
+        response = client.post(
+            "/plugins/upload/", files={"file": (zip_file_name, f, "application/zip")}
+        )
+
+    # request was processed
+    assert response.status_code == 200
+    assert response.json()["filename"] == zip_file_name
+
+    ### each test function having `just_installed_plugin` as argument, is run here
+    yield
+    ###
+
+    # clean up of zip file and mock_plugin_folder is done for every test automatically (see client fixture)
